\name{.cleanFilename}
\alias{.cleanFilename}
\title{This function removes spaces and punctuations from filenames.}
\usage{
  .cleanFilename(x)
}
\arguments{
  \item{x}{filename}
}
\value{
  filename
}
\description{
  This function removes spaces and punctuations from
  filenames.
}
\keyword{internal}

