\name{convertFiles}
\alias{convertFiles}
\title{This function converts mass spectra stored in fid/mzXML files into
ASCII/CSV/MSD files.}
\usage{
  convertFiles(path, from = c("fid", "mzXML"),
    to = c("ascii", "csv", "msd"))
}
\arguments{
  \item{path}{path to files}

  \item{from}{source file format}

  \item{to}{target file format}
}
\description{
  This function converts mass spectra stored in fid/mzXML
  files into ASCII/CSV/MSD files.
}

