\docType{methods}
\name{fwhm}
\alias{fwhm}
\alias{fwhm,list,list-method}
\alias{fwhm,MassSpectrum,MassPeaks-method}
\title{FWHM}
\arguments{
  \item{spec}{a \code{\linkS4class{MassSpectrum}} object or
  a list of \code{\linkS4class{MassSpectrum}} objects}

  \item{peaks}{a \code{\linkS4class{MassPeaks}} object or a
  list of \code{\linkS4class{MassPeaks}} objects}
}
\description{
  This method calculates the \emph{f}ull-\emph{w}idth at
  \emph{h}alf-\emph{m}aximum for each peak.
}
\seealso{
  \code{\linkS4class{MassPeaks}},
  \code{\linkS4class{MassSpectrum}}
}
\keyword{methods}

