\docType{package}
\name{MALDIquantTools-package}
\alias{MALDIquantTools}
\alias{MALDIquantTools-package}
\alias{package-MALDIquantTools}
\title{Additional tools for MALDIquant}
\description{
  Additional tools for MALDIquant. This package is not
  intended for public release on CRAN.
}
\details{
  \tabular{ll}{ Package: \tab MALDIquantTools \cr Version:
  \tab 0.5.2\cr Date: \tab 2012-08-14\cr License: \tab GPL
  (>= 3)\cr URL: \tab
  http://strimmerlab.org/software/maldiquant/\cr }
}
\author{
  Sebastian Gibb <\email{mail@sebastiangibb.de}>
}
\references{
  \url{http://strimmerlab.org/software/maldiquant/}
}
\keyword{package}

