\docType{data}
\name{uniProtTable}
\alias{uniProtTable}
\title{UniProt reference table}
\description{
  This dataset contains mass for UniProt listed proteins.
}
\details{
  To create: \preformatted{ uniProtTable <-
  MALDIquantTools:::createUniProtReferenceTable(organism=9606,
  keyword=181, reviewed=TRUE) }
}
\references{
  \url{http://www.uniprot.org/}
}
\seealso{
  \code{\link[MALDIquantTools]{averagineTable}},
  \code{\link[MALDIquantTools]{createUniProtReferenceTable}}
}
\keyword{data}
\keyword{datasets}

