\name{.useBRAIN}
\alias{.useBRAIN}
\title{This function is a wrapper around \code{\link[BRAIN]{useBRAIN}}.}
\usage{
  .useBRAIN(aC, nRatio = 3)
}
\arguments{
  \item{aC}{list with fields C, H, N, O, S of integer
  non-negative values.}

  \item{nRatio}{number of ratio products.}
}
\value{
  data.frame
}
\description{
  This function is a wrapper around
  \code{\link[BRAIN]{useBRAIN}}.
}
\seealso{
  \code{\link[BRAIN]{useBRAIN}}
}
\keyword{internal}

